requirements.txt
flask
python-nmap
openai==0.28
pyyaml
requests
dnspython